+++ src/lib.rs
@@ -41,11 +41,12 @@ pub mod prelude {
     pub use super::physics::backend::rapier::TiledPhysicsRapierBackend;
     #[cfg(feature = "physics")]
     pub use super::physics::{
-        backend::{multi_polygon_as_line_strings, multi_polygon_as_triangles, TiledPhysicsBackend},
+        backend::TiledPhysicsBackend,
         collider::{
             ColliderCreated, TiledColliderOf, TiledColliderPolygons, TiledColliderSource,
             TiledColliders,
         },
+        geometry::{multi_polygon_as_line_strings, multi_polygon_as_triangles, simplify_geometry},
         settings::TiledPhysicsSettings,
         TiledPhysicsPlugin,
     };
@@ -61,15 +62,15 @@ pub mod prelude {
             grid_size_from_map, tile_size, tilemap_type_from_map,
         },
         image::TiledImage,
-        layer::{TiledLayer, TiledLayerParallax, TiledParallaxCamera},
+        layer::{TiledLayer, TiledLayerId, TiledLayerParallax, TiledParallaxCamera},
         map::{
             asset::TiledMapAsset, loader::TiledMapLoaderError, storage::TiledMapStorage,
             RespawnTiledMap, TiledMap, TiledMapImageRepeatMargin, TiledMapLayerZOffset,
             TiledMapReference,
         },
-        object::{TiledObject, TiledObjectVisualOf, TiledObjectVisuals},
+        object::{TiledObject, TiledObjectId, TiledObjectVisualOf, TiledObjectVisuals},
         sets::{TiledPostUpdateSystems, TiledPreUpdateSystems, TiledUpdateSystems},
-        tile::{TiledTile, TiledTilemap},
+        tile::{TiledTile, TiledTileId, TiledTilemap, TiledTilesetId},
         world::{
             asset::TiledWorldAsset, chunking::TiledWorldChunking, loader::TiledWorldLoaderError,
             storage::TiledWorldStorage, RespawnTiledWorld, TiledWorld,